v1.0.30 & v1.0.29 - Execute psudo feature (#2)
* 1.0.29 Execute support

Execute vanilla command now supports psudo commands. The old behavior with only psudo still works, except for psudoas. The equivalent to "psudoas <target> <command>" is "execute as <target> run psudoas <command>".

With execute, selectors (e.g. @s) and coordinates (e.g. for distance=) from any psudo depends now on the last "as" of the execute command. To make someone run the command after psudo, this should be psudoas and not psudo and this person must be selected in the last "as" of execute command.

Don't forget to use psudouuid with mobs to replace the selector by the uuid and not a poor "pig", even with @s !

(Based on the other fork https://github.com/Numa-Lab/PsudoCommands)

* 1.0.29 (end) Support tab completion again

* 1.0.29 (end) Add behavior for execute "at"

* Update changelog.txt

* Update README.md

* Fix when sender is CommandBlock,  fix local coords

* Add command psudoasop

The new command psudoasop is like psudoas except for plugin commands. If the following command is a plugin command, the pre-check permissions will be ignored, i.e. the permission set to the command in the plugin.yml is ignored so the player can perform the command, so the command DOES NOT give operator permission. If the plugin command checks permissions inside onCommand method, then psudoasop will still fail if the sender hasn't the permissions and I can't do anything for this, developers should know how to link permissions to a command.

Be careful with psudoasop, for now timings don't work.

* Add support for spigot server from 1.14.4 to 1.19

* Execute psudo feature (#1)

* 1.0.29 Execute support

Execute vanilla command now supports psudo commands. The old behavior with only psudo still works, except for psudoas. The equivalent to "psudoas <target> <command>" is "execute as <target> run psudoas <command>".

With execute, selectors (e.g. @s) and coordinates (e.g. for distance=) from any psudo depends now on the last "as" of the execute command. To make someone run the command after psudo, this should be psudoas and not psudo and this person must be selected in the last "as" of execute command.

Don't forget to use psudouuid with mobs to replace the selector by the uuid and not a poor "pig", even with @s !

(Based on the other fork https://github.com/Numa-Lab/PsudoCommands)

* Support tab completion again

* Add behavior for execute "at"

* Update changelog.txt

* Fix when sender is CommandBlock,  fix local coords

* Add command psudoasop

The new command psudoasop is like psudoas except for plugin commands. If the following command is a plugin command, the pre-check permissions will be ignored, i.e. the permission set to the command in the plugin.yml is ignored so the player can perform the command, so the command DOES NOT give operator permission. If the plugin command checks permissions inside onCommand method, then psudoasop will still fail if the sender hasn't the permissions and I can't do anything for this, developers should know how to link permissions to a command.

Be careful with psudoasop, for now timings don't work.

* Add support for spigot server from 1.14.4 to 1.19

* 1.0.30 Add timings when using Paper (get NoSuchFieldException when using Spigot... wtf)